+++ CHANGELOG.md
@@ -1,7 +1,26 @@
-## 0.0.1
+## 0.0.6
 
-* Create Flutter number picker.
+* Update info, fix errors.
+* Custom add, minus button.
+
+## 0.0.5
+
+* Update info, fix errors.
+
+## 0.0.4
+
+* Update info, fix errors.
+
+## 0.0.3
+
+* Update info, fix errors.
 
 ## 0.0.2
 
 * Update info, fix errors.
+
+## 0.0.1
+
+* Create Flutter number picker.
+
+